Move Histograms and Scalars under metrics package with common MetricDefinition ancestor
