+++ src/main/scala/com/mozilla/telemetry/metrics/Histogram.scala
@@ -1,4 +1,4 @@
-package com.mozilla.telemetry.histograms
+package com.mozilla.telemetry.metrics
 
 import org.json4s._
 import org.json4s.jackson.JsonMethods._
@@ -7,7 +7,7 @@ import scala.io.Source
 
 case class RawHistogram(values: Map[String, Int], sum: Long)
 
-sealed abstract class HistogramDefinition
+sealed abstract class HistogramDefinition extends MetricDefinition
 case class FlagHistogram(keyed: Boolean) extends HistogramDefinition
 case class BooleanHistogram(keyed: Boolean) extends HistogramDefinition
 case class CountHistogram(keyed: Boolean) extends HistogramDefinition
