+++ src/main/scala/com/mozilla/telemetry/metrics/MetricDefinition.scala
@@ -1,0 +1,5 @@
+package com.mozilla.telemetry.metrics
+
+trait MetricDefinition {
+  val keyed: Boolean
+}
