+++ src/main/scala/com/mozilla/telemetry/metrics/Scalars.scala
@@ -1,4 +1,4 @@
-package com.mozilla.telemetry.scalars
+package com.mozilla.telemetry.metrics
 
 import java.util
 
@@ -6,7 +6,7 @@ import org.yaml.snakeyaml.Yaml
 import collection.JavaConversions._
 import scala.io.Source
 
-sealed abstract class ScalarDefinition
+sealed abstract class ScalarDefinition extends MetricDefinition
 case class UintScalar(keyed: Boolean) extends ScalarDefinition
 case class BooleanScalar(keyed: Boolean) extends ScalarDefinition
 case class StringScalar(keyed: Boolean) extends ScalarDefinition
