+++ src/test/scala/com/mozilla/telemetry/LongitudinalTest.scala
@@ -2,8 +2,7 @@ package com.mozilla.telemetry.views
 import java.io.PrintWriter
 
 import com.mozilla.telemetry.parquet.ParquetFile
-import com.mozilla.telemetry.scalars.ScalarsClass
-import com.mozilla.telemetry.histograms.HistogramsClass
+import com.mozilla.telemetry.metrics.{ScalarsClass, HistogramsClass}
 import org.apache.avro.Schema
 import org.apache.avro.generic.GenericRecord
 import org.apache.spark.sql.{Row, SQLContext}
